Confidence Interval for an Unstandardized Regression Coefficient

Computes a confidence interval for a population regression coefficient in its raw (unstandardized) metric, by the standard t-based approach or the noncentral t approach. A thin convenience wrapper around ci_reg_coef, which is the general engine; for the standardized coefficient use ci_src.

Usage

ci_rc(
  b_j,
  SE_b_j = NULL,
  s_Y = NULL,
  s_X = NULL,
  N,
  p,
  R2_Y_X = NULL,
  R2_j_X_without_j = NULL,
  conf_level = 0.95,
  R2_Y_X_without_j = NULL,
  t_value = NULL,
  alpha_lower = NULL,
  alpha_upper = NULL,
  noncentral = FALSE,
  ...
)

Arguments

b_j

Value of the regression coefficient for the jth predictor variable

SE_b_j

Standard error for the jth predictor variable

s_Y

standard Deviation of Y, the dependent variable

s_X

standard Deviation of X, the predictor variable of interest

N

Sample size

p

The number of predictors

R2_Y_X

The squared multiple correlation coefficient predicting Y from the p predictor variables

R2_j_X_without_j

The squared multiple correlation coefficient predicting the jth predictor variable (i.e., the predictor of interest) from the remaining p-1 predictor variables

conf_level

Desired level of confidence for the computed interval (i.e., 1 - the Type I error rate)

R2_Y_X_without_j

The squared multiple correlation coefficient predicting Y from the p-1 predictor variable with the jth predictor of interest excluded

t_value

The t-value evaluating the null hypothesis that the population regression coefficient for the jth predictor equals zero

alpha_lower

The Type I error rate for the lower confidence interval limit

alpha_upper

The Type I error rate for the upper confidence interval limit

noncentral

TRUE or FALSE statement specifying whether or not the noncentral approach to confidence intervals should be used

...

Optional additional specifications for nested functions

Value

A data.frame with rows for the lower and upper confidence limits of the regression coefficient, returned as a single value column keyed by term.

Details

Returns the confidence limits for the regression coefficient of interest from the standard approach to confidence interval formation or from the noncentral approach to confidence interval formation using the noncentral t-distribution.

Note

Not all of the values need to be specified, only those that contain all of the necessary information in order to compute the confidence interval (options are thus given for the values that need to be specified).

References

Kelley, K. (2007). Confidence intervals for standardized effect sizes: Theory, application, and implementation. Journal of Statistical Software, 20(8), 1–24. doi:10.18637/jss.v020.i08

Kelley, K., & Maxwell, S. E. (2003). Sample size for multiple regression: Obtaining regression coefficients that are accurate, not simply significant. Psychological Methods, 8(3), 305–321. doi:10.1037/1082-989X.8.3.305

Kelley, K., & Maxwell, S. E. (2008). Sample size planning with applications to multiple regression: Power and accuracy for omnibus and targeted effects. In P. Alasuutari, L. Bickman, & J. Brannen (Eds.), The Sage handbook of social research methods (pp. 166–192). Sage.

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2027). Designing experiments and analyzing data: A model comparison perspective (4th ed.). Routledge. (See Chapter 4 on individual comparisons of means and Chapter 6 on trend analysis.)

Smithson, M. (2003). Confidence intervals. New York, NY: Sage Publications.

Steiger, J. H. (2004). Beyond the F Test: Effect size confidence intervals and tests of close fit in the Analysis of Variance and Contrast Analysis. Psychological Methods, 9(2), 164–182. doi:10.1037/1082-989X.9.2.164

See also

Author

Ken Kelley kkelley@nd.edu

Examples

ci_rc(b_j = 0.61319, SE_b_j = 0.16098, N = 30, p = 6, conf_level = 0.95)
#>  term        value prob_less prob_greater
#>  lower_limit 0.28  0.025     0.975       
#>  upper_limit 0.946 0.975     0.025       
#> 
#> Confidence level: 95%