Conversion Functions for Noncentral t-distribution

Functions useful for converting a standardized mean difference to a noncentrality parameter, and vice versa.

Usage

convert_delta_lambda(delta, n_1, n_2)

convert_lambda_delta(lambda, n_1, n_2)

Arguments

delta

Population value of the standardized mean difference

n_1

Sample size in group 1

n_2

Sample size in group 2

lambda

noncentral value from a t-distribution

Value

Each function returns a 1-row data.frame with columns term and value. The term entry identifies the conversion ("delta_lambda" or "lambda_delta") and value is the converted scalar. The two functions are exact inverses given the per-group sample sizes n_1 and n_2.

Details

Although lambda is the population noncentral value, an estimate of it is the observed value of a t-statistic. Likewise, delta can be estimated as the observed standardized mean difference. Thus, the observed standardized mean difference can be converted to the observed t-value. These functions are especially helpful in the context of forming confidence intervals for the population standardized mean difference.

References

Kelley, K. (2007). Confidence intervals for standardized effect sizes: Theory, application, and implementation. Journal of Statistical Software, 20(8), 1–24. doi:10.18637/jss.v020.i08

See also

Author

Ken Kelley kkelley@nd.edu

Examples

convert_lambda_delta(lambda = 2, n_1 = 113, n_2 = 113)
#>  term         value
#>  lambda_delta 0.266
convert_delta_lambda(delta = .266076, n_1 = 113, n_2 = 113)
#>  term         value
#>  delta_lambda 2